cont refactor
