+++ app/ante/evm/10_increment_sequence.go
@@ -1,0 +1,66 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
+package evm
+
+import (
+	errorsmod "cosmossdk.io/errors"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
+
+	evmtypes "github.com/evmos/evmos/v15/x/evm/types"
+
+	"github.com/ethereum/go-ethereum/common"
+)
+
+// EthIncrementSenderSequenceDecorator increments the sequence of the signers.
+type EthIncrementSenderSequenceDecorator struct {
+	ak evmtypes.AccountKeeper
+}
+
+// NewEthIncrementSenderSequenceDecorator creates a new EthIncrementSenderSequenceDecorator.
+func NewEthIncrementSenderSequenceDecorator(ak evmtypes.AccountKeeper) EthIncrementSenderSequenceDecorator {
+	return EthIncrementSenderSequenceDecorator{
+		ak: ak,
+	}
+}
+
+// AnteHandle handles incrementing the sequence of the signer (i.e. sender). If the transaction is a
+// contract creation, the nonce will be incremented during the transaction execution and not within
+// this AnteHandler decorator.
+func (issd EthIncrementSenderSequenceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (sdk.Context, error) {
+	for _, msg := range tx.GetMsgs() {
+		_, txData, from, err := evmtypes.UnpackEthMsg(msg)
+		if err != nil {
+			return ctx, err
+		}
+
+		// increase sequence of sender
+		acc := issd.ak.GetAccount(ctx, from)
+		if acc == nil {
+			return ctx, errorsmod.Wrapf(
+				errortypes.ErrUnknownAddress,
+				"account %s is nil", common.BytesToAddress(from.Bytes()),
+			)
+		}
+		nonce := acc.GetSequence()
+
+		// we merged the nonce verification to nonce increment, so when tx includes multiple messages
+		// with same sender, they'll be accepted.
+		if txData.GetNonce() != nonce {
+			return ctx, errorsmod.Wrapf(
+				errortypes.ErrInvalidSequence,
+				"invalid nonce; got %d, expected %d", txData.GetNonce(), nonce,
+			)
+		}
+
+		if err := acc.SetSequence(nonce + 1); err != nil {
+			return ctx, errorsmod.Wrapf(err, "failed to set sequence to %d", acc.GetSequence()+1)
+		}
+
+		issd.ak.SetAccount(ctx, acc)
+	}
+
+	return next(ctx, tx, simulate)
+}
